requirements.txt
numpy
scipy
transforms3d
pyquaternion
pyserial
opencv-python
rospkg
catkin-tools
networkx
paho-mqtt
pymavlink
utm
witmotion
digi-xbee